src/main/java/tn/esprit/spring/kaddem/entities/EquipeDTO.java
package tn.esprit.spring.kaddem.entities;

public class EquipeDTO {
    private Integer idEquipe;
    private String nomEquipe;
    private String niveau;

    // Getters and Setters
    public Integer getIdEquipe() {
        return idEquipe;
    }

    public void setIdEquipe(Integer idEquipe) {
        this.idEquipe = idEquipe;
    }

    public String getNomEquipe() {
        return nomEquipe;
    }

    public void setNomEquipe(String nomEquipe) {
        this.nomEquipe = nomEquipe;
    }

    public String getNiveau() {
        return niveau;
    }

    public void setNiveau(String niveau) {
        this.niveau = niveau;
    }
}
